requirements.txt
streamlit>=1.40.0
pandas>=2.2.3
python-dotenv>=1.0.1
requests>=2.32.3
deep-translator>=1.11.4
chromadb>=1.4.0
sentence-transformers>=3.0.1
numpy>=1.26.4
playwright>=1.49.0
openpyxl>=3.1.5
